refactor(metrics): remove the inert metrics-api ngrx actions
MetricsAPIAction/MetricsStratosAction were dispatched from the metrics
page and the endpoint-details row, but their effect+reducer were deleted
in W-a4 (2847befc2b) and MetricsAPIActionSuccess was never consumed by
any reducer even before that. The dispatches have always been no-ops and
metrics_targets/metrics_stratos are never populated by anything, so
removal is a zero-runtime-change cleanup.

Delete the action classes, action constants and the unused
MetricAPIResponse/MetricAPIQueryTypes; relocate the still-referenced
MetricsAPITargets/MetricsStratosInfo metadata shapes to
types/metrics-api.types.ts. Drop the two dispatches and the now-unused
Store injection from both metrics components.
