+++ src/frontend/packages/store/src/public-api.ts
@@ -87,8 +87,7 @@ export { MetricQueryType } from './types/metric.types';
 export { CATALOGUE_ENTITIES, EntityCatalogFeatureModule, EntityCatalogModule } from './entity-catalog.module';
 export { EntityCatalogProvidersModule } from './entity-catalog-providers.module';
 export { EntityServiceFactory } from './entity-service-factory.service';
-export type { MetricsAPITargets } from './actions/metrics-api.actions';
-export { MetricsAPIAction, MetricsStratosAction } from './actions/metrics-api.actions';
+export type { MetricsAPITargets, MetricsStratosInfo } from './types/metrics-api.types';
 export type { ListView } from './actions/list.actions';
 export { ListFilter, ListPagination, ListSort, SetListViewAction } from './actions/list.actions';
 export { getActions } from './actions/action.helper';
